requirements.txt
Django==3.1.1
pandas
djangorestframework
django-cors-headers
scipy